Product entity fields updated

+++ src/main/java/com/canngos/shoppingbasketservice/entity/Product.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -15,11 +17,11 @@ public class Product {
     @Column(unique = true, nullable = false)
     private String name;
     @Column(nullable = false)
-    private Double price;
+    private BigDecimal price;
     private String description;
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
     private ProductType type = ProductType.OTHER;
     @Column(nullable = false)
-    private int quantity;
+    private Integer quantity;
 }
